Multiple changes

+++ app/components/EventBox.jsx
@@ -3,18 +3,34 @@ import {Link, IndexLink} from 'react-router';
 
 var EventBox = React.createClass({
   deleteEvent: function() {
-    this.props.deleteEvent(this.props.eventKey);
+    this.getConfirmation();
+  },
+  getConfirmation: function(){
+     var retVal = confirm("Are you sure you want to delete this event?");
+     if( retVal == true ){
+       this.props.deleteEvent(this.props.eventKey);
+     }
   },
   render: function () {
-    var {eventKey, id, name, date, time} = this.props;
+    var {eventKey, id, name, date, time, totalRevenue, totalExpenses, net} = this.props;
 
     return (
       <div className="event-box">
-        <span onClick={this.deleteEvent} className="delete-x">X</span>
+        <span onClick={this.getConfirmation} className="delete-x">X</span>
         <Link to={"events/"+ eventKey} className="event-box-link">
-            <p>{name}</p>
-            <p>{date}</p>
-            <p>{time}</p>
+          <div className="row">
+            <div className="large-6 columns">
+              <p><strong>{name}</strong></p>
+              <p>{date}</p>
+              <p>{time}</p>
+            </div>
+            <div className="large-6 columns">
+              <p>Revenue: ${totalRevenue}</p>
+              <p>Expenses: ${totalExpenses}</p>
+              <p>Net: ${net}</p>
+            </div>
+          </div>
+
         </Link>
       </div>
